requirements.txt
# python 3.11
streamlit
tensorflow
torch
torchvision
ultralytics
Pillow